+++ docs/history.rst
@@ -10,6 +10,7 @@ History and release notes
 * Added :meth:`Parser.optional`.
 * Added :meth:`Parser.tag`.
 * Added :func:`seq` keyword argument version (Python 3.6)
+* Added :meth:`Parser.combine_dict`.
 * Documentation improvements.